mvctp_v2 updated.

+++ MVCTP/SenderStatusProxy.cpp
@@ -258,18 +258,18 @@ void SenderStatusProxy::TcpTransferFile(string file_name) {
 
 
 // Generate a local data file for disk-to-disk transfer experiments
-int SenderStatusProxy::GenerateDataFile(string file_name, unsigned long bytes) {
+int SenderStatusProxy::GenerateDataFile(string file_name, ulong bytes) {
 	int buf_size = 256;
 	char buffer[buf_size];
 	for (int i = 0; i < buf_size; i++) {
 		buffer[i] = i;
 	}
 
-	unsigned long remained_size = bytes;
+	ulong remained_size = bytes;
 	ofstream myfile(file_name.c_str(), ios::out | ios::trunc);
 	if (myfile.is_open()) {
 		while (remained_size > 0) {
-			int len = remained_size < buf_size ? remained_size : buf_size;
+			ulong len = remained_size < buf_size ? remained_size : buf_size;
 			myfile.write(buffer, len);
 			remained_size -= len;
 		}
